+++ freemarker-generator-base/src/main/java/org/apache/freemarker/generator/base/uri/NamedUriStringParser.java
@@ -33,7 +33,7 @@ import static java.util.regex.Pattern.compile;
  *     <li>file:///users.csv</li>
  *     <li>users=file:///users.csv</li>
  *     <li>users:admin=file:///users.csv</li>
- *     <li>users=file:///users.csv#charset=UTF-16&amp;mimetype=text/csv</li>
+ *     <li>users=file:///users.csv#charset=UTF-16&amp;mimeType=text/csv</li>
  * </ul>
  */
 public class NamedUriStringParser {
@@ -52,10 +52,10 @@ public class NamedUriStringParser {
         if (matcher.matches()) {
             final String name = matcher.group(NAME);
             final String group = matcher.group(GROUP);
-            final URI uri = UriUtils.toURI(matcher.group(URI));
+            final URI uri = UriUtils.toUri(matcher.group(URI));
             return new NamedUri(name, group, uri, parameters(uri));
         } else {
-            final URI uri = UriUtils.toURI(value);
+            final URI uri = UriUtils.toUri(value);
             return new NamedUri(uri, parameters(uri));
         }
     }