+++ freemarker-generator-base/src/main/java/org/apache/freemarker/generator/base/util/StringUtils.java
@@ -51,5 +51,4 @@ public class StringUtils {
         }
         return count;
     }
-
 }