+++ freemarker-generator-base/src/main/java/org/apache/freemarker/generator/base/util/UriUtils.java
@@ -16,17 +16,16 @@
  */
 package org.apache.freemarker.generator.base.util;
 
-import org.apache.commons.io.FilenameUtils;
-
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.net.URL;
 
+import static org.apache.commons.io.FilenameUtils.separatorsToUnix;
 import static org.apache.freemarker.generator.base.util.StringUtils.isNotEmpty;
 
 public class UriUtils {
 
-    public static URI toURI(String str) {
+    public static URI toUri(String str) {
         try {
             return new URI(separatorsToUnix(str));
         } catch (URISyntaxException e) {
@@ -34,19 +33,31 @@ public class UriUtils {
         }
     }
 
-    public static URI toURI(String scheme, String path) {
-        return toURI(scheme + ":///" + path);
+    public static URI toUri(String scheme, String path) {
+        return toUri(scheme + ":///" + path);
     }
 
-    public static URI toURI(URL url) {
-        return toURI(url.toString());
+    public static URI toUri(URL url) {
+        return toUri(url.toString());
+    }
+
+    /**
+     * Return the URI string representation without fragment part.
+     *
+     * @param uri uri
+     * @return string representation of URI without fragment part
+     */
+    public static String toStringWithoutFragment(URI uri) {
+        final String str = uri.toString();
+        final int index = str.indexOf('#');
+        return (index > 0) ? str.substring(0, index) : str;
     }
 
     public static boolean isUri(String str) {
         return isNotEmpty(str) && str.contains("://");
     }
 
-    public static boolean isHttpURI(URI uri) {
+    public static boolean isHttpUri(URI uri) {
         if (uri == null) {
             return false;
         }
@@ -65,9 +76,5 @@ public class UriUtils {
             return false;
         }
         return "env".equalsIgnoreCase(uri.getScheme());
-    }
-
-    private static String separatorsToUnix(String str) {
-        return FilenameUtils.separatorsToUnix(str);
     }
 }